Decoupled Input from fighter
Added an input component that catches input from controllers. This means
any object can be controlled by the player, CPU, network, or just random
garbage as long as a controller is written for it. The current player
controller basically rips the input straight from the input manager and
sends it along.